Add wh handler back to test if bug resolves with migratiosn applied
