+++ checkout/urls.py
@@ -1,5 +1,6 @@
 from django.urls import path
 from . import views
+from .webhooks import webhook
 
 urlpatterns = [
     path('', views.checkout, name='checkout'),
@@ -7,4 +8,6 @@ urlpatterns = [
     path('cart/', views.view_cart, name='view_cart'),
     path('add/<item_id>', views.add_to_cart, name='add_to_cart'),
     path('remove/<item_id>/', views.remove_from_cart, name='remove_from_cart'),
+    path('cache_checkout_data/', views.cache_checkout_data, name='cache_checkout_data'),
+     path('wh/', webhook, name='webhook'),
 ]
