+++ checkout/webhook_handlerCOPY.py
@@ -1,0 +1,44 @@
+"""
+Webhook handlers
+"""
+
+import json
+import time
+from django.http import HttpResponse
+from django.conf import settings
+from services.models import Service
+from .models import Order, OrderItem
+
+
+class StripeWH_Handler:
+    """Handle Stripe webhooks"""
+
+    def __init__(self, request):
+        self.request = request
+
+    def handle_event(self, event):
+        """
+        Handle a generic/unknown/unexpected webhook event
+        """
+        return HttpResponse(
+            content=f'Unhandled Webhook received: {event["type"]}',
+            status=200)
+
+    def handle_payment_intent_succeeded(self, event):
+        """
+        Handle the payment_intent.succeeded webhook from Stripe
+        """
+        intent = event.data.object
+        
+        return HttpResponse(
+            content=f'Webhook received: {event["type"]}',
+            status=200)
+
+
+    def handle_payment_intent_payment_failed(self, event):
+        """
+        Handle the payment_intent.payment_failed webhook from Stripe
+        """
+        return HttpResponse(
+            content=f'Webhook received: {event["type"]}',
+            status=200)
