+++ help_my_plants/settings.py
@@ -138,7 +138,6 @@ else:
 
 
 
-
 # Password validation
 # https://docs.djangoproject.com/en/3.2/ref/settings/#auth-password-validators
